Moved source files to src/ subdir. Removed debug messages.
