--- contains.h
@@ -1,52 +1,0 @@
-/** @file
-    @brief Header
-
-    @date 2015
-
-    @author
-    Sensics, Inc.
-    <http://sensics.com>
-
-*/
-
-// Copyright 2015 Sensics, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-
-#ifndef INCLUDED_contains_h_GUID_AB217A8C_822E_40AC_B3AA_9BDEF9563EC0
-#define INCLUDED_contains_h_GUID_AB217A8C_822E_40AC_B3AA_9BDEF9563EC0
-
-// Internal Includes
-// - none
-
-// Library/third-party includes
-// - none
-
-// Standard includes
-#include <algorithm>        // for std::find
-#include <iterator>         // for std::begin, std::end
-
-/**
- * Searches a container for an element.
- *
- * @returns true if the element is in the countainer, false otherwise.
- */
-template <typename T, typename E>
-bool contains(const T& container, const E& element)
-{
-    return (std::find(std::begin(container), std::end(container), element) != std::end(container));
-}
-
-#endif // INCLUDED_contains_h_GUID_AB217A8C_822E_40AC_B3AA_9BDEF9563EC0
-
