+++ src/OculusRiftException.h
